Refacto: add instollowing class (#2)
* refacto: add instollowing class

This commit adds a new class called "Instollowing" to the code. The class has several methods for parsing JSON files, creating output folders and files, and generating HTML files. It also includes a method for getting followers and following from Instagram. The main function in "instollowing.py" now creates an instance of the "Instollowing" class and calls its "run" method.

* ci: add format python

* chore: delete unuseful files

* chore: delete unuseful files

* Apply Black Formatting

* ci: edit the ci

---------

Co-authored-by: github-actions <github-actions@github.com>