ci:Add Codecov workflow for code coverage
This commit adds a new file `.github/workflows/code-cov.yml` which contains the configuration for running code coverage checks using Codecov. The workflow is triggered on push and pull requests to the `master` branch. It runs on `ubuntu-latest` and includes a step to upload coverage reports to Codecov using the `codecov/codecov-action@v3` action. The necessary environment variable `CODECOV_TOKEN` is set using secrets.